+++ packages/onekijs-framework/src/auth/LoginService.ts
@@ -13,7 +13,7 @@ import { sha256 } from '../utils/crypt';
 import { get } from '../utils/object';
 import { absoluteUrl } from '../utils/router';
 import AuthService from './AuthService';
-import { LoginState, Mfa } from './typings';
+import { Idp, LoginState, Mfa } from './typings';
 import {
   generateCodeChallenge,
   generateCodeVerifier,
@@ -325,19 +325,52 @@ export default class LoginService extends DefaultLocalService<LoginState> {
         [token, securityContext] = yield callback(response, idp, this.context);
       }
 
-      if (isOauth(idp) && idp.nonce && get(token as any, 'id_token')) {
-        // validates the nonce found in the id_token
-        // https://openid.net/specs/openid-connect-core-1_0.html#TokenResponseValidation
-        const id_token = parseJwt((token as { id_token: string }).id_token);
-        const nonce = getIdpStorage(idp).getItem(`onekijs.${identity}.nonce`);
-        const hash: string = yield sha256(nonce || undefined);
-        getIdpStorage(idp).removeItem(`onekijs.${identity}.nonce`);
-        if (hash !== id_token.nonce) {
-          throw Error('Invalid oauth2 nonce');
+
+      let mfa: Mfa = { required: false };
+      if (typeof idp.mfa === 'function') {
+        yield idp.mfa(response, idp, this.context);
+      } else if (idp.mfa) {
+        mfa = idp.mfa;
+      } else {
+        mfa = {
+          required: get(token, 'mfa_required', false),
+          token: get(token, 'mfa_token'),
+          token_type: get(token, 'token_type'),
+          totpSecret: get(token, 'totp_secret'),
+          user: get(token, 'mfa_user'),
+          delay: get(token, 'mfa_delay'),
         }
       }
 
-      yield this.successLogin(token as any, securityContext, idpName, onError, onSuccess);
+      if (mfa.required) {
+        // we have to add an additional step (ask the TOTP to the user)
+        // if totp_secret is set, it means that the user haven't yet configured his MFA
+
+        const noMfaToken = this.authService.loadNoMfaToken(idp, identity);
+
+        if (noMfaToken) {
+          yield this.sendNoMfaToken(noMfaToken, mfa, idp, onSuccess);
+        } else {
+          yield this.setMfa(mfa);
+          yield this.onSuccess();
+        }
+      } else {
+        if (isOauth(idp) && idp.nonce && get(token as any, 'id_token')) {
+          // validates the nonce found in the id_token
+          // https://openid.net/specs/openid-connect-core-1_0.html#TokenResponseValidation
+          const id_token = parseJwt((token as { id_token: string }).id_token);
+          const nonce = getIdpStorage(idp).getItem(`onekijs.${identity}.nonce`);
+          const hash: string = yield sha256(nonce || undefined);
+          getIdpStorage(idp).removeItem(`onekijs.${identity}.nonce`);
+          if (hash !== id_token.nonce) {
+            throw Error('Invalid oauth2 nonce');
+          }
+        }
+
+        yield this.successLogin(token as any, securityContext, idpName, onError, onSuccess);
+      }
+
+
     } catch (e) {
       if (process.env.NODE_ENV === 'development') {
         console.error('External login callback error', e);
@@ -431,6 +464,7 @@ export default class LoginService extends DefaultLocalService<LoginState> {
           token_type: get(response, 'token_type'),
           totpSecret: get(response, 'totp_secret'),
           user: data.username,
+          delay: get(response, 'mfa_delay'),
         }
       }
 
@@ -441,7 +475,14 @@ export default class LoginService extends DefaultLocalService<LoginState> {
         }
         // we have to add an additional step (ask the TOTP to the user)
         // if totp_secret is set, it means that the user haven't yet configured his MFA
-        yield this.setMfa(mfa);
+        const noMfaToken = this.authService.loadNoMfaToken(idp, identity);
+        if (noMfaToken) {
+          yield this.sendNoMfaToken(noMfaToken, mfa, idp, onSuccess);
+        } else {
+          yield this.setMfa(mfa);
+          yield this.onSuccess();
+        }
+
       } else {
         yield this.formLoginResponse(response, idpName, onError, onSuccess);
       }
@@ -462,6 +503,22 @@ export default class LoginService extends DefaultLocalService<LoginState> {
       }
     }
     yield this.setLoading(false);
+  }
+
+  @saga(SagaEffect.Latest)
+  *sendNoMfaToken(noMfaToken: string, mfa: Mfa, idp: Idp, onSuccess?: SuccessCallback) {
+    try {
+      yield this.verifyTotp({
+        no_mfa_token: noMfaToken,
+        token: mfa.token,
+        token_type: mfa.token_type,
+      }, idp.name, undefined, onSuccess);
+    } catch (e) {
+      // error during the usage of the no_mfa_token, just ask the verification code to the user
+      yield this.setMfa(mfa);
+      yield this.onSuccess();
+    }
+
   }
 
 /**
@@ -535,11 +592,18 @@ export default class LoginService extends DefaultLocalService<LoginState> {
           throw new DefaultBasicError(`Invalid totpEndpoint for IDP ${idp.name}`);
         }
 
-        const body = {
-          totp: data.totp,
-        };
+        const body: AnonymousObject = {}
+        if (data.totp) {
+          body.totp = data.totp;
+        }
+        if (data.trusted_device) {
+          body.trusted_device = data.trusted_device;
+        }
+        if (data.no_mfa_token) {
+          body.no_mfa_token = data.no_mfa_token;
+        }
 
-        const token_type = get(this.state.mfa, 'token_type');
+        const token_type = data.token_type ?? get(this.state.mfa, 'token_type');
 
         response = yield asyncHttp(absoluteUrl(url, get(settings, 'server.baseUrl')), 'POST', body, {
           headers: {